requirements.txt
langchain_chroma==1.1.0
langchain_community==0.4.1
langchain_core==1.2.23
langchain_huggingface==1.2.1
langchain_openai==1.1.12
pydantic==2.12.5
python-dotenv==1.2.2
